# app-backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Babble API",
    "version": "1.0.0",
    "description": "Babble API — 발음 연습 서비스"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          },
          "errorKey": {
            "type": "string"
          }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {}
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          },
          "errorKey": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/assessments": {
      "post": {
        "tags": [
          "진단(Assessment)"
        ],
        "summary": "음성 파일 업로드 및 진단 생성",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "audio": {
                    "type": "string",
                    "format": "binary"
                  },
                  "scriptId": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "업로드 성공 및 진단 생성 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "파일이 없거나 잘못된 형식",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "진단(Assessment)"
        ],
        "summary": "진단 이력 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/assessments/trial/policy": {
      "get": {
        "tags": [
          "진단(Assessment)"
        ],
        "summary": "게스트 체험 정책 조회",
        "description": "게스트 기능별 최대 체험 횟수와 Assessment TTL 정보를 반환한다. 인증 불필요.",
        "responses": {
          "200": {
            "description": "정책 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "maxTrials": {
                          "type": "object",
                          "description": "기능별 최대 체험 횟수",
                          "properties": {
                            "PRACTICE": {
                              "type": "integer",
                              "example": 1
                            },
                            "CONTINUOUS_READING": {
                              "type": "integer",
                              "example": 1
                            },
                            "WORD_GAME": {
                              "type": "integer",
                              "example": 0
                            },
                            "LEADERBOARD": {
                              "type": "integer",
                              "example": 0
                            },
                            "HISTORY": {
                              "type": "integer",
                              "example": 0
                            },
                            "PROFILE": {
                              "type": "integer",
                              "example": 0
                            }
                          }
                        },
                        "assessmentTtlHours": {
                          "type": "integer",
                          "example": 24
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit 초과"
          }
        }
      }
    },
    "/assessments/notifications/sse": {
      "get": {
        "tags": [
          "진단(Assessment)"
        ],
        "summary": "실시간 분석 알림 SSE (일반 사용자)",
        "description": "사용자의 진단 분석 완료/실패를 실시간으로 수신하는 SSE 스트림",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "SSE 스트림 연결 성공",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "503": {
            "description": "연결 제한 초과"
          }
        }
      }
    },
    "/assessments/admin/user/{userId}": {
      "get": {
        "tags": [
          "진단 관리"
        ],
        "summary": "(어드민) 특정 사용자의 전체 진단 이력 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "관리자 권한 필요"
          }
        }
      }
    },
    "/assessments/admin/all": {
      "get": {
        "tags": [
          "진단 관리"
        ],
        "summary": "(어드민) 모든 사용자의 전체 진단 이력 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "관리자 권한 필요"
          }
        }
      }
    },
    "/assessments/admin/{id}/retry": {
      "post": {
        "tags": [
          "진단 관리"
        ],
        "summary": "(어드민) 진단 분석 재시도",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "재시도 요청 성공"
          }
        }
      }
    },
    "/assessments/admin/notifications/sse": {
      "get": {
        "tags": [
          "진단 관리"
        ],
        "summary": "(어드민) 실시간 분석 알림 SSE",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "SSE 연결 성공"
          }
        }
      }
    },
    "/assessments/{id}": {
      "get": {
        "tags": [
          "진단(Assessment)"
        ],
        "summary": "특정 진단 결과 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공"
          },
          "404": {
            "description": "진단 결과를 찾을 수 없음"
          }
        }
      }
    },
    "/assessments/{id}/retry": {
      "post": {
        "tags": [
          "진단(Assessment)"
        ],
        "summary": "진단 분석 재시도",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "재시도 요청 성공"
          },
          "404": {
            "description": "진단 결과를 찾을 수 없음"
          }
        }
      }
    },
...
  },
...
}